requirements-dev.txt
pytest
pytest-asyncio
pytest-homeassistant-custom-component
coverage
hacs
